% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_eval.R
\name{cor_eval}
\alias{cor_eval}
\title{A tool to evaluate and plot correlation structure}
\usage{
cor_eval(dat, cormeth = "spearman", labsize = 1, varlabs = NULL)
}
\arguments{
\item{dat}{is a data object}

\item{cormeth}{passes correlation method to cor()}

\item{labsize}{sets label size}

\item{varlabs}{specifies variable labels}
}
\value{
a correlation matrix
}
\description{
cor_eval explores data correlations
}
\examples{
#NIEHS Mixtures Workshop dataset1
data(dataset1)
cor_eval(dataset1[,2:9])
}
