% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumcount.r
\name{cumcount}
\alias{cumcount}
\title{Get cumulative count}
\usage{
cumcount(x)
}
\arguments{
\item{x}{A vector for which to get cumulative count}
}
\value{
The cumulative count of all items in x
}
\description{
Get the cumulative count of a variable of interest
}
