% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpad.r
\name{lagpad}
\alias{lagpad}
\title{Lag a vector}
\usage{
lagpad(x, k = 1)
}
\arguments{
\item{x}{A vector to be lagged}

\item{k}{The number of lags to output}
}
\value{
The lagged vector with NA's in missing values
}
\description{
Create a vector of the lag of a variable and fill missing values with NA's.
}
