% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ecm.R
\name{map_ecm}
\alias{map_ecm}
\title{map_ecm: Constructs an Exposure Continuum Map using Kohonen's Self-Organizing Map algorithm}
\usage{
map_ecm(
  trn_dat,
  xdim = 5,
  ydim = 4,
  maptopo = NULL,
  itermax = NULL,
  inits = NULL,
  distmet = NULL,
  lmode = NULL
)
}
\arguments{
\item{trn_dat}{is a data object to map. Should be a numerical or factor based data matrix.}

\item{xdim}{is x-dimension of the SOM}

\item{ydim}{is y-dimension of the SOM}

\item{maptopo}{specifies the topology of the SOM grid as either "rectangular" or "hexagonal"}

\item{itermax}{specifies the number if learning iterations}

\item{inits}{specifies if optimal initialization values are to be used.}

\item{distmet}{specifies dissimilarity metric. Current options include "sumofsquares", "euclidean", "manhattan", and "tanimoto". Default is to use "Euclidean" for continuous data, and "tanimoto" for factors.}

\item{lmode}{specifies the learning algorithm. The default is "online" but "batch" and "pbatch" are available via kohonen}
}
\value{
a 'kohonen' object
}
\description{
This function applies Kohonen's Self-Organizing Map algorithm in order to construct a low-dimensional mapping of the data.
}
\references{
Kohonen, T. (1995) Self-Organizing Maps. Springer-Verlag
}
