% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_inits.R
\name{map_inits}
\alias{map_inits}
\title{map_inits: Identifies initial values for optimal Exposure Continuum Map fit via SOM}
\usage{
map_inits(
  trn_dat,
  xdim = 4,
  ydim = 3,
  maptopo = NULL,
  itermax = NULL,
  seedopt = "pca.sample",
  seedeval = "KL",
  distmet = NULL,
  nstarts = NULL,
  lmode = NULL
)
}
\arguments{
\item{trn_dat}{is a data object to map. Should be a numerical or factor based data matrix.}

\item{xdim}{is x-dimension of the SOM}

\item{ydim}{is y-dimension of the SOM}

\item{maptopo}{specifies the topology of the SOM grid as either "rectangular" or "hexagonal"}

\item{itermax}{specifies the number if learning iterations}

\item{seedopt}{specifies method for initialization values. "pca.sample", "pca" or "ran" accepted.}

\item{seedeval}{specifies the evaluation statistic for seed optimalization.}

\item{distmet}{specifies dissimilarity metric. Current options include "sumofsquares", "euclidean", "manhattan", and "tanimoto". Default is to use "Euclidean" for continuous data, and "tanimoto" for factors.}

\item{nstarts}{specifies number of initialization schemes to test if inits are not provided}

\item{lmode}{specifies the learning algorithm. The default is "online" but "batch" and "pbatch" are available via kohonen}
}
\value{
a data.frame containing evaluation measures
}
\description{
map_inits: Identifies initial values for optimal Exposure Continuum Map fit via SOM
}
