% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metric.R
\name{map_metric}
\alias{map_metric}
\title{A function for extracting classifications and coordinates from self-Organizing Map (SOM) models}
\usage{
map_metric(map_obj)
}
\arguments{
\item{map_obj}{an object with class 'kohonen'}
}
\value{
a dataframe object with classiciation and coordinate assignments for each training observation.
}
\description{
map_stats calculates common measures of model stats for a SOM model
}
