% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Plots profiles using radial segment diagrams}
\usage{
map_plot(
  map_obj,
  varnames = NULL,
  colormod = NULL,
  nodelab = TRUE,
  labsize = 1,
  labtype = "IDs",
  addXY = TRUE,
  addFreq = FALSE,
  freqtype = "frq",
  freqsize = 1,
  addleg = "TRUE",
  legloc = "bottom",
  legsymsize = 2,
  leglabsize = 1,
  legtxtlas = 2,
  ...
)
}
\arguments{
\item{map_obj}{a SOM object}

\item{varnames}{specifies variable names}

\item{colormod}{can be used to customize plot color schemes. The default is terrain.colors().}

\item{nodelab}{whether or not to display reference labels on each node}

\item{labsize}{sets size of node labels}

\item{labtype}{specifies type of label. "IDs" or "XYs" accepted.}

\item{addXY}{adds map coordinates to plot}

\item{addFreq}{whether or not to display frequency labels}

\item{freqtype}{determines type of frequency labels. Relative frequencies can be specified with "frq" or counts with "cnts"}

\item{freqsize}{frequency label cex value}

\item{addleg}{adds legend to map}

\item{legloc}{defaults to "bottom" but "right" is also option}

\item{legsymsize}{legend symbol cex value}

\item{leglabsize}{legend text cex value}

\item{legtxtlas}{legend text position rotation passed to las. Numeric in {0,1,2,3}; the style of axis label.}

\item{...}{other parameters}
}
\value{
a kohonen map where radial segment diagrams are used to illustrate profiles for each map node
}
\description{
Plots profiles using radial segment diagrams
}
\details{
Radial segment diagrams provide a compact way to illustrate profiles derived from SOM. Here diagrams are presented along the SOM grid in order provide a compact visualization of SOM results.s the map.
}
