% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stats.R
\name{map_stats}
\alias{map_stats}
\title{A collection of measures for evaluating Self-Organizing Maps (SOM) models}
\usage{
map_stats(map_obj)
}
\arguments{
\item{map_obj}{an object with class 'kohonen'}
}
\value{
a data.frame of values

TSS     The total sum of squares

WCSS    The within-class sum of squares

TOTWCSS The total within-class sum of squares

BCSS    The between-class sum of squares, i.e. TSS-TOTWCSS

R2     The total variance explained by the SOM classes (BCSS/TSS)

ADJ_R2 The total variance explained by the SOM classes adjusted for class number

MAE     The mean absolute error (MAE) derived from the mean class assignment distance, i.e., quantization error

RMSE    The root-mean-square-error derived from class assignment distances, i.e., quantization error
}
\description{
map_stats calculates common measures of model stats for a SOM model
}
