% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_eval.R
\name{var_eval}
\alias{var_eval}
\title{A data assessment and screening tool}
\usage{
var_eval(dat, labsize = 1, symsize = 2, colormod = NULL, varlabs = NULL)
}
\arguments{
\item{dat}{is a data frame object}

\item{labsize}{specifies cex value of labels on plots default is 0.75}

\item{symsize}{specifies cex value of symbols on plots default is 2}

\item{colormod}{can be used to customize plot color schemes. The default uses terrain palletes from grDevices}

\item{varlabs}{specifies x-axis variable labels}
}
\value{
a dataframe consisting of summary statistics
\itemize{
\item {VARIABLE} Name of variable
\item {TYPE}  variable class
\item {N}  number of observations
\item {NAs}  number of missing values
\item {ZEROs}  number of zero values
\item {MEAN}  mean value
\item {SD}  standard deviation
\item {MIN}  minimum value
\item {Q1}  lower quartile
\item {MED}  median value
\item {Q3}  upper quartile
\item {MAX}  maximum value
\item {IqR}  interquartile range
\item {CoV}  coefficient of variation, i.e., SD/MEAN
\item {CVrank}  ranking of CoV values
}
}
\description{
Performs a generalized evaluation of the data in order to examine standardized distributions of data via boxplots, variable specific missing-value patterns, and variable specific zero-value patterns. A table of summary statistics is also provided.
}
\details{
An important first step in the evaluation of multivariate data sets is to examine candidate variables individually. This assessment and screening tool is designed to assist understanding individual data distributions and patterning with problematic features such as missing and zero values.
}
\examples{
#NIEHS Mixtures Workshop dataset1
data(dataset1)
var_eval(dataset1)
}
